+++ src/graphics/light/PointLight.hpp
@@ -1,15 +1,16 @@
 #ifndef POINTLIGHT_H
 #define POINTLIGHT_H
 
-#include "../../lib/glm/glm.hpp"
+#include "Light.hpp"
+//~ #include "../../lib/glm/glm.hpp"
 
 class PointLight : public Light
 {
 	private:
-		glm::vec3 position;
+		glm::dvec3 position;
 	public:
-		void setPosition( glm::vec3 position ){ this->position = position; }
-		glm::vec3 getPosition(){ return position; }
+		void setPosition( glm::dvec3 position ){ this->position = position; }
+		glm::dvec3 getPosition(){ return position; }
 };
 
 #endif // POINTLIGHT_H
