+++ src/graphics/light/SpotLight.hpp
@@ -1,21 +1,22 @@
 #ifndef SPOTLIGHT_H
 #define SPOTLIGHT_H
 
-#include "../../lib/glm/glm.hpp"
+#include "Light.hpp"
+//~ #include "../../lib/glm/glm.hpp"
 
 class SpotLight : public Light
 {
 	private:
-		glm::vec3 position;
-		glm::vec3 direction;
+		glm::dvec3 position;
+		glm::dvec3 direction;
 		double falloffAlpha1;
 		double falloffAlpha2;
 		
 	public:
-		void setPosition( glm::vec3 position ){ this->position = position; }
-		glm::vec3 getPosition(){ return position; }
-		void setDirection( glm::vec3 direction ){ this->direction = direction; }
-		glm::vec3 getDirection(){ return direction; }
+		void setPosition( glm::dvec3 position ){ this->position = position; }
+		glm::dvec3 getPosition(){ return position; }
+		void setDirection( glm::dvec3 direction ){ this->direction = direction; }
+		glm::dvec3 getDirection(){ return direction; }
 		void setFalloffAlpha1( double falloffAlpha1 ){ this->falloffAlpha1 = falloffAlpha1; }
 		double getFalloffAlpha1(){ return falloffAlpha1; }
 		void setFalloffAlpha2( double falloffAlpha2 ){ this->falloffAlpha2 = falloffAlpha2; }
