+++ src/graphics/material/MaterialSolid.hpp
@@ -2,15 +2,14 @@
 #define MATERIALSOLID_H
 
 #include "Material.hpp"
-#include "../../lib/glm/glm.hpp"
 
 class MaterialSolid : public Material
 {
 	private:
-		glm::vec3 color;
+		glm::dvec3 color;
 	public:
-		void setColor( glm::vec3 color ){ this->color = color; }
-		virtual glm::vec3 getColor(){ return color; }
+		void setColor( glm::dvec3 color ){ this->color = color; }
+		virtual glm::dvec3 getColor(){ return color; }
 };
