+++ src/graphics/material/MaterialTextured.hpp
@@ -3,14 +3,13 @@
 
 #include "Material.hpp"
 #include "string"
-#include "../../lib/glm/glm.hpp"
 
 class MaterialTextured : public Material
 {
 	private:
 		std::string texture;
 	public:
-		virtual glm::vec3 getColor(){ return glm::vec3(0,0,0); }
+		virtual glm::dvec3 getColor(){ return glm::dvec3(0,0,0); }
 };
 
 #endif // MATERIALTEXTURED_H
