+++ src/graphics/surface/Surface.cpp
@@ -3,12 +3,12 @@
 #include "../../lib/glm/gtx/string_cast.hpp"
 //~ #include <string>
 
-void Surface::translate( glm::vec3 translate )
+void Surface::translate( glm::dvec3 translate )
 {
 	transformations = glm::translate( transformations, translate );
 } 
 
-void Surface::scale( glm::vec3 scale )
+void Surface::scale( glm::dvec3 scale )
 {
 	transformations = glm::scale( transformations, scale );
 }
@@ -16,17 +16,17 @@ void Surface::scale( glm::vec3 scale )
 // important glm::rotate uses degrees and only accepts float
 void Surface::rotateX( double rotateX )
 {
-	transformations = glm::rotate( transformations, (float)rotateX, glm::vec3(1,0,0) );
+	transformations = glm::rotate( transformations, rotateX, glm::dvec3(1,0,0) );
 }
 
 void Surface::rotateY( double rotateY )
 {
-	transformations = glm::rotate( transformations, (float)rotateY, glm::vec3(0,1,0) );
+	transformations = glm::rotate( transformations, rotateY, glm::dvec3(0,1,0) );
 }
 
 void Surface::rotateZ( double rotateZ )
 {
-	transformations = glm::rotate( transformations, (float)rotateZ, glm::vec3(0,0,1) );
+	transformations = glm::rotate( transformations, rotateZ, glm::dvec3(0,0,1) );
 }
 
 std::string Surface::transformationMatrixString()
