+++ src/io/PPMWriter.hpp
@@ -14,7 +14,7 @@ class PPMWriter
 	public:
 		void setResolution( int horizontal, int vertical );
 		void setFilename( std::string filename ){ this->filename = filename; }
-		void writePPM( glm::vec3* image );
+		void writePPM( glm::dvec3* image );
 };
 
 inline void PPMWriter::setResolution( int horizontal, int vertical )
